retinonet_backend/requirements.txt
dnspython==2.7.0
fastapi==0.115.12
gdown==5.2.0
mediapipe==0.10.21
numpy==2.2.4
opencv_contrib_python==4.11.0.86
opencv_python==4.8.1.78
Pillow==11.1.0
python-dotenv==1.0.1
reportlab==4.2.5
tensorflow==2.15.0
tensorflow_intel==2.15.0
gunicorn